Removed the need for traversal functions to return anything
In hind-sight this made no sense. The traversal functions are called repeatedly
for each item on the list, having a return value is not possible since each function
can potentially return something different.

Currently, no application of the list requires any return value during traversal. If
this becomes a requirement then the application should create a specalised traversal
function to handle this case.