Fixed bug with checking if list is empty
Previously when checking if list is empty it was assumed
that the list object was not set to NULL. This is a bug
since the deleteList function explicitly sets the list to
NULL after deleting every element.

This was a necessary fix to ensure the counting of the list
elements works correctly because a check for an empty list
is done first.